El sistema ahora permite la creación, consulta y movimiento de fondos entre cuentas corrientes con integridad de datos garantizada y comunicación estable con MySQL

+++ src/main/java/bank/application/inputs/SavingsAccountService.java
@@ -5,9 +5,9 @@ import bank.domain.SavingsAccount;
 import java.math.BigDecimal;
 
 public interface SavingsAccountService {
-    void withdraw(String accountNumber, double amount);
+    void withdraw(String accountNumber, BigDecimal amount);
     void applyInterest(String accountNumber);
-    void deposit(String accountNumber, double amount);
+    void deposit(String accountNumber, BigDecimal amount);
     SavingsAccount getAccount(String accountNumber);
 
     // Esta es la firma que debe coincidir con tu clase:
